+++ wire/cli/src/cli.rs
@@ -31,9 +31,9 @@ pub struct Cli {
     #[command(flatten)]
     pub verbose: clap_verbosity_flag::Verbosity<WarnLevel>,
 
-    /// Path to directory containing hive
-    #[arg(long, global = true, default_value = std::env::current_dir().unwrap().into_os_string())]
-    pub path: std::path::PathBuf,
+    /// Path or flake reference
+    #[arg(long, global = true, default_value = std::env::current_dir().unwrap().into_os_string(), visible_alias("flake"))]
+    pub path: String,
 
     // Unused until a solution to tracing-indicatif log deadlocking is found...
     /// Hide progress bars.