+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `--ssh-accept-host` was added.
 - `--on -` will now read additional apply targets from stdin.
 - `{key.name}-key.{path,service}` systemd units where added.
+- `--path` now supports flakerefs (`github:foo/bar`, `git+file:///...`,
+  `gitlab:foo/bar`, etc).
+- `--flake` is now an alias for `--path`.
 
 ### Fixed